requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
mlx-lm
sentencepiece==0.2.0
llama-cpp-python
debugpy